Revert "Turn controller refactoring"
